requirements.txt
matplotlib==3.7.1
Pillow==9.5.0
Pillow==10.2.0
tqdm==4.65.0
